SLG_website/requirements.txt
asgiref==3.5.2
coverage==6.4.3
Django==4.0.5
django-debug-toolbar==3.5.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
mysqlclient==2.1.1
Pillow==9.2.0
sqlparse==0.4.2
tzdata==2022.1
